requirements.txt
yandex-cloud-ml-sdk
python-dotenv